mScopeParsers/mScopeTransform/requirements.txt
pyparsing==2.2.0
python-dateutil==2.6.0
six==1.10.0
mScopeTransformer==0.1dev

#git+https://github.com/coc-gatech-newelba/NewElbaAlpha.git@master#egg=mScopeTransformer&subdirectory=mScopeParsers/mScopeTransform
